backend/requirements.txt
# FastAPI and web framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.27
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Redis and caching
redis==5.0.1
hiredis==2.3.2

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8

# Email
aiosmtplib==3.0.2
email-validator==2.1.0

# Telephony (for Telegram bot)
python-telegram-bot==20.7
aiohttp==3.9.1

# AI and ML
google-generativeai==0.3.0
scikit-learn==1.4.0
pandas==2.1.4
numpy==1.26.3

# Celery for async tasks
celery==5.3.6
flower==2.0.1

# Reports and exports
openpyxl==3.1.2
reportlab==4.0.7

# HTTP requests
httpx~=0.25.2  # Compatible with python-telegram-bot 20.7

# Rate limiting
slowapi==0.1.9

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0

# Utilities
python-dateutil==2.8.2
pytz==2024.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0

# Development
black==24.1.1
flake8==7.0.0
mypy==1.8.0

